+++ src/main/java/com/dubbi/blogplatform/application/dto/OAuthEnrollDto.java
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 public class OAuthEnrollDto {
     private String email;
-    private int age;
+    private long age;
     private String city;
 }
